refactor(GameSetting): Changed name (Computer -> GameSetting)
