+++ README.md
@@ -4,10 +4,9 @@
 ### 게임 플로우
 자동차 생성 및 이름 부여 -> 몇번 이동 할지 결정 -> 게임 진행 -> 우승자 판별 
 ### 필요 기능
-##### 자동차 n대 이름 입력 기능
+##### 자동차 n대 이름 입력 바탕으로 자동차 생성 기능
 - 예외: 이름이 5자 이하가 아님 / 중복된 이름
 - "," 기준으로 입력 (e.g. pobi,woni,jun)
-##### 입력 바탕으로 자동차 생성 기능
 ##### 이동 횟수 입력 기능
 - 예외: 정수가 아님
 ##### 게임 시작 기능
