--- src/main/java/racingcar/Announcer.java
@@ -1,4 +1,0 @@
-package racingcar;
-
-public class Announcer {
-}
